Basic implementation of SGE
Added AbstractProblemSGE with the same functionality as AbstractProblemGE but with arrays of Integers, StructuredGramaticalEvolution that extends SimpleGeneticAlgorithm and uses its functions and ProblemSGE_example to test the implementation.

BnfReader has been changed to add more functionalities, but I might separete them in another class in the future.

FInally added the operators, UniformCrossover and IntegerFlipMutationList.

Issue with cloning and arrays since the cloning method in Variable won't make a clone of the array and therefore the arrays will be shared by different individuals, the solution that is implemented now is to make a new array in the mutation since it is the only operation that changes values inside the list and  when not changing the inside values just leave the pointers of different objects to point to the same place.